Bootstrapping and core feature of a pastebin service.
Just go to the site root, then type the code you want in the page and click "submit". This redirects the user to a url to share so that it can be reused at will.The codes are stored in ram, so they are volatile. Use of the python bottle library for the backend, and the hashlib library to easily create a link. On the Javascript side, the highlight.js library is used to provide syntax highlighting on the codes.

Co-authored-by: Dr Lazor <julien.castiaux@gmail.com>