fix invalid form charset
Upload a bin with non-ascii characters like the following French
sentence: "Le maître va à l'école". When saved the letters `î`, `à` and
`é` are not displayed right.

A charset is the rules used to map stream bytes to text characters,
bottle uses Latin-1, a charset defining 256 characters on 2-bytes when
it reads the html form it receives. This encoding is popular in western
Europe as it maps all the characters used in the many spoken language
and offers some additional mathematic symbols. The encoding is also
popular in engineering as it is a reversible.

In Latin-1, there are 256 characters encoded on 1-byte, every single
character is mapped to exactly one byte and every byte is mapped to
excatly one character. It is possible to re-encode the text to its
byte stream representation and to re-decode it using the proper charset.

Fixes #3